pyproject.toml
[tool.poetry]
name = "consumer-ai-hackathon-london"
version = "0.1.0"
description = ""
authors = ["Matt <matthewrharding@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.1.1"
fastapi = "^0.115.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
